clean up includefile a bit, remove from_env (not used)
